[removal] merge deprecated CampaignScheduledEvent into ScheduledEvent
